requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.0.0
cryptography==36.0.2
decorator==5.1.1
dm-tree==0.1.7
flatbuffers==2.0
frozenlist==1.3.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
keras==2.7.0
Keras-Preprocessing==1.1.2
Markdown==3.3.6
multidict==6.0.2
numpy==1.21.6
oauthlib==3.2.0
opt-einsum==3.3.0
pandas==1.3.5
pip==22.0.4
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pyu2f==0.1.5
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.3
setuptools==62.1.0
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-probability==0.15.0
termcolor==1.1.0
typing_extensions==4.2.0
umap==0.1.1
urllib3==1.26.9
Werkzeug==2.1.1
wheel==0.37.1
wrapt==1.14.0
yarl==1.7.2
zipp==3.8.0
